huggingface/agent_course/unit_2_smolagent/pyproject.toml
[project]
name = "unit-2-smolagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0", # required by the DuckDuckGoSearchTool
    # "huggingface-hub>=0.36.0",
    "python-dotenv>=1.2.1",
    "smolagents>=1.23.0",
]
